requirements.txt
docker-py>=0.2.2
flask>=0.10